refactor: move server util to api-server module
