+++ src/utils/api-client.ts
@@ -1,32 +1,4 @@
-import { getServerSession } from 'next-auth/next'
-
-import { options as authOptions } from '@/app/api/auth/options'
 import type { Post } from '@/types/blog'
-export const accessPermitted = async () => {
-  const session = await getServerSession(authOptions)
-
-  //early return if user is not logged in
-  if (!session) {
-    return {
-      permitted: false,
-      response: new Response('You must be logged in.', {
-        status: 401,
-      }),
-    }
-  }
-
-  //early return if user is not admin
-  if (session.user.role !== 'admin') {
-    return {
-      permitted: false,
-      response: new Response('Only admin can create a blog post.', {
-        status: 403,
-      }),
-    }
-  }
-
-  return { permitted: true, response: null }
-}
 
 export const updatePostClientSide = async (id: number, data: Post) => {
   const response = await fetch(
